fixed bug in applier

+++ driver.py
@@ -34,7 +34,6 @@ def main(user, secret, search):
         print('Error: No resumes. You have either not added a resume or the resume is corrupted or deleted')
         return
     
-
     driver = login(user, secret, search)
     if driver == None:
         print("Error: login failed")
@@ -44,7 +43,7 @@ def main(user, secret, search):
     
     # main loop
     main = driver.window_handles[0]
-    wait = WebDriverWait(driver, 10)
+    wait = WebDriverWait(driver, 60)
     bottons = wait.until(EC.presence_of_all_elements_located((By.CLASS_NAME,'btn-primary')))
 
     for btn in bottons:
@@ -54,35 +53,38 @@ def main(user, secret, search):
             driver.switch_to.window(driver.window_handles[1])
             time.sleep(0.5)
             title = scraper(driver)
-
-            driver.switch_to.window(driver.window_handles[-1])
-            cover_letter, resume = selector(title)
-            flag = createCoverLetter(title, cover_letter)
-            upload_cover_letter(driver, title + '.pdf')
-            upload_resume(driver, resume)
-
-            #print('Apply online?', flag)
-            if flag:
-                todo = os.path.join(current, 'TODO', title)
-                cover_path = os.path.join(cover_dr, title + '.pdf')
-                resume_path = os.path.join(resume_dr, resume)
-                shutil.copy2(cover_path, todo)
-                shutil.copy2(resume_path, todo)
-
-            driver.switch_to.window(driver.window_handles[1])
-            applier(driver, title)
-            driver.close()
-            driver.switch_to.window(main)
             
-            # delete created cover letters
-            
-            if os.path.exists(os.path.join(cover_dr, title + '.docx')):
+            if title != "NONE":
+                driver.switch_to.window(driver.window_handles[-1])
+                cover_letter, resume = selector(title)
+                flag = createCoverLetter(title, cover_letter)
+                upload_cover_letter(driver, title + '.pdf')
+                upload_resume(driver, resume)
+
+                #print('Apply online?', flag)
+                if flag:
+                    todo = os.path.join(current, 'TODO', title)
+                    cover_path = os.path.join(cover_dr, title + '.pdf')
+                    resume_path = os.path.join(resume_dr, resume)
+                    shutil.copy2(cover_path, todo)
+                    shutil.copy2(resume_path, todo)
+
+                driver.switch_to.window(driver.window_handles[1])
+                applier(driver, title)
+                driver.close()
+                driver.switch_to.window(main)
                 
-                os.remove(os.path.join(cover_dr, title + '.docx'))
-            
-            if os.path.exists(os.path.join(cover_dr, title + '.pdf')):
-                os.remove(os.path.join(cover_dr, title + '.pdf'))
-            
+                # delete created cover letters
+                
+                if os.path.exists(os.path.join(cover_dr, title + '.docx')):
+                    
+                    os.remove(os.path.join(cover_dr, title + '.docx'))
+                
+                if os.path.exists(os.path.join(cover_dr, title + '.pdf')):
+                    os.remove(os.path.join(cover_dr, title + '.pdf'))
+            else:
+                driver.switch_to.window(main)
+
             
     
     time.sleep(20)
